+++ R/archive.R
@@ -13,8 +13,8 @@
 #'
 #' @return For read operations: the ichimoku object originally archived.
 #'
-#'     For write operations: invisible NULL. 'object' is written to
-#'     'filename' as a side effect.
+#'     For write operations: invisible NULL. 'object' is written to 'filename'
+#'     as a side effect.
 #'
 #' @details For read operations, please specify only 'filename'. 'filename' is
 #'     read and the return value may be assigned to an object.
@@ -23,14 +23,21 @@
 #'     'object' will be written to 'filename'. Confirmation is printed to the
 #'     console if the file write operation has been successful.
 #'
+#' @section Further Details:
+#'
+#'     This function requires the 'arrow' package to be installed.
+#'
+#'     If the 'openssl' package is available, a sha256 hash of the original
+#'     object is written to the archive. This allows the data integrity of the
+#'     restored object to be verified when the archive is read back.
+#'
 #' @examples
 #' cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
-#'
 #' filename <- tempfile()
+#'
 #' archive(cloud, filename)
 #'
 #' restored <- archive(filename)
-#' identical(cloud, restored)
 #'
 #' unlink(filename)
 #'
@@ -71,7 +78,7 @@ archive <- function(..., object, filename) {
     } else stop("in archive(object, filename): 'object' specified without 'filename'", call. = FALSE)
 
   } else {
-    message("Note: please install the 'arrow' package to enable archiving of ichimoku objects",
+    message("Please install the 'arrow' package to enable archiving of ichimoku objects",
             "\nArchives utilise the Apache Arrow IPC file format")
   }
 }
@@ -100,16 +107,34 @@ writeArchive <- function(object, filename) {
          "', archive() only writes ichimoku objects", call. = FALSE)
   }
 
+  if (file.exists(filename)) {
+    continue <- readline(prompt = paste0("The file '", filename,
+                                         "' already exists. Overwrite? [y/N] "))
+    if (!continue %in% c("y", "Y", "yes", "YES")) {
+      message("Request cancelled")
+      return(invisible())
+    }
+  }
+
+  if (requireNamespace("openssl", quietly = TRUE)) {
+    sha256 <- suppressWarnings(openssl::sha256(as.raw(object)))
+  } else {
+    message("sha256 hash not written to archive as 'openssl' package not available - ",
+            "\nthis means the integrity of objects cannot be verified when read back")
+    sha256 <- NULL
+  }
   df <- xts_df(object)
   df <- structure(df,
-                  ichimoku = TRUE,
                   periods = attr(object, "periods"),
                   periodicity = attr(object, "periodicity"),
                   ticker = attr(object, "ticker"),
-                  strat = attr(object, "strat"))
+                  strat = attr(object, "strat"),
+                  ichimoku351 = TRUE,
+                  sha256 = sha256)
 
   arrow::write_feather(df, filename)
-  message("Archive created: ", filename)
+  message("Archive written to '", filename, "'", sep = "")
+  invisible()
 }
 
 #' Read ichimoku objects from Arrow Archive
@@ -132,17 +157,39 @@ readArchive <- function(filename) {
   }
 
   df <- arrow::read_feather(filename)
-  if (!isTRUE(attr(df, "ichimoku"))) {
+  if (!isTRUE(attr(df, "ichimoku351"))) {
     stop("'", paste0(filename), "' is not an archived ichimoku object", call. = FALSE)
   }
 
   tzone <- if (is.null(attr(df$index, "tzone"))) "" else attr(df$index, "tzone")
 
-  structure(xts(df[, -1], order.by = df[, 1], tzone = tzone),
-            class = c("ichimoku", "xts", "zoo"),
-            periods = attr(df, "periods"),
-            periodicity = attr(df, "periodicity"),
-            ticker = attr(df, "ticker"),
-            strat = attr(df, "strat"))
+  object <- structure(xts(df[, -1], order.by = df[, 1], tzone = tzone),
+                      class = c("ichimoku", "xts", "zoo"),
+                      periods = attr(df, "periods"),
+                      periodicity = attr(df, "periodicity"),
+                      ticker = attr(df, "ticker"),
+                      strat = attr(df, "strat"))
+
+  if (is.null(attr(df, "sha256"))) {
+    message("Archive read from '", filename,
+            "'\nIntegrity of restored object cannot be verified as archive does not contain a sha256 hash")
+
+  } else if (requireNamespace("openssl", quietly = TRUE)) {
+    sha256 <- suppressWarnings(openssl::sha256(as.raw(object)))
+    if (identical(sha256, attr(df, "sha256"))) {
+      message("Archive read from '", filename,
+              "'\nData verified by sha256: ", sha256)
+    } else {
+      message("Archive read from '", filename, "'")
+      warning("sha256 of restored object\n", sha256,
+              " does not match the original\n", attr(df, "sha256"), call. = FALSE)
+    }
+  } else {
+    message("Archive read from '", filename,
+            "'\n'openssl' package required for integrity of restored objects to be verified")
+  }
+
+  object
+
 }